rework(database): improve service interface
- rename `StateService` to `StateDatabase`
- now service API can be scoped to workspace by calling
  `state_db.workspace(id).workspace_related_call(...)`
- move `EditShelf`, `EditWorkspace`, `CreateTag` and `DeleteTag`
  workspace-related logic inside `StateDatabase`
